In-Class/Labs/3 - JUnit/IntelliJ Project/Lab3-JUnit/src/main/java/org/example/Employee.java
package org.example;

public class Employee {

    private String name;
    private int id;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        if(id > 0) {
            this.id = id;
        }
    }

    private double salary;

    public double getSalary() {
        return salary;
    }

    public void setSalary(double salary) {
        if(salary < 100 || salary > 1) {
            this.salary = salary;
        }
    }

}
